back/services/auth_server/requirements.txt
Django
django-environ
djangorestframework-simplejwt
social-auth-app-django
psycopg2-binary
gunicorn
python-decouple
dj-database-url
djangorestframework
django-cors-headers